Use package `io.joern.joerncli` for project `joerncli`
