+++ joern-cli/src/main/scala/io/joern/joerncli/console/Joern.scala
@@ -1,3 +1,3 @@
-package io.joern.console
+package io.joern.joerncli.console
 
 object Joern extends JoernConsole {}
