+++ joern-cli/src/main/scala/io/joern/joerncli/console/JoernAmmoniteExecutor.scala
@@ -1,4 +1,4 @@
-package io.joern.console
+package io.joern.joerncli.console
 
 import io.joern.console.scripting.AmmoniteExecutor
