+++ joern-cli/src/test/scala/io/joern/joerncli/GenerationTests.scala
@@ -1,10 +1,9 @@
-package io.joern
+package io.joern.joerncli
 
 import better.files.File
-import org.scalatest.wordspec.AnyWordSpec
-import org.scalatest.matchers.should.Matchers
-
 import io.shiftleft.semanticcpg.language._
+import org.scalatest.matchers.should.Matchers
+import org.scalatest.wordspec.AnyWordSpec
 
 /** Test code that shows how code property graphs can be generated using the FuzzyC language frontend
   */
